requirements.txt
python-dotenv~=1.0.0
supabase~=1.0.3
streamlit~=1.25.0
elevenlabs~=0.2.21
pandas~=2.0.3
langchain~=0.0.240